Do not modify third-party keymaps when loading this library
Merely loading a library should not modify the keymaps of other
packages, as this may override bindings added by the user.

One way to do it in style would be to provide a global minor mode,
which enables minor modes in the appropriate major modes, and use the
keymaps of these minor modes to add the new key bindings.  But that is
a lot of work and is less flexible than to just instruct users to add
the necessary bindings, while suggesting (but not enforcing) certain
keys.